+++ source/relation/table.h
@@ -11,7 +11,7 @@ SelectManager *relation_table_select_manager(RelationTable *table);
 
 SelectManager *relation_table_project(RelationTable *table, SqlLiteral sql_literal);
 
-SelectManager *relation_table_limit(RelationTable *table, int limit_number);
+SelectManager *relation_table_limit(RelationTable *table, long long limit_number);
 
 SelectManager *relation_table_where(RelationTable *table, char *expression);
